rework: Move getAll to a helper
