+++ server/.env.example
@@ -15,7 +15,8 @@ HASH_AMOUNT=10
 ENABLE_LOGIN=true
 ENABLE_REGISTER=true
 
-# Pagination - NOTE: To match the UI, should be divisible by 3
+# Customization
 PAGINATION=9
+CODE_LENGTH=4
 
 # vim: ft=sh
